// challenge/hybris/bin/ext-smartedit/cmssmartedit/web/features/cmssmarteditContainer/components/navigation/types.ts
/*
 * [y] hybris Platform
 *
 * Copyright (c) 2018 SAP SE or an SAP affiliate company. All rights reserved.
 *
 * This software is the confidential and proprietary information of SAP
 * ("Confidential Information"). You shall not disclose such Confidential
 * Information and shall use it only in accordance with the terms of the
 * license agreement you entered into with SAP.
 */
import {TreeNode, TypedMap} from "smarteditcommons";

export interface NavigationNodeEntry {
	id: string;
	name: string;
	itemType: string;
}

export interface NavigationNode extends TreeNode {
	id: string;
	nodes: NavigationNode[];
	entries?: NavigationNodeEntry[];
	parent: NavigationNode;
}

export interface NavigationNodeCMSItem {
	catalogVersion: string;
	visible: boolean;
	title: TypedMap<string>;
	uuid: string;
	uid: string;
	entries: string[];
	pages: number[];
	itemtype: string;
	modifiedtime: string;
	children: string[];
	name: string;
	links: string[];
	creationtime: string;
}